Fixing WCF Service to Work on localhost again
For now on when adding a new Service Configuration just build a new Project to save time on any backtracking. There goes 3 hours of my life for a pointless service thingy
