--- WCFCalculatorService/Service.svc
@@ -1,1 +1,0 @@
-﻿<%@ ServiceHost Language="C#" Debug="true" Service="CalculatorService" CodeBehind="~/App_Code/CalculatorService.cs" %>
