--- WCFCalculatorService/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Runtime" version="4.3.1" targetFramework="net472" />
-</packages>
