+++ WCFCalculatorServiceLocal/Web.config
@@ -1,17 +1,19 @@
 ﻿<?xml version="1.0"?>
 <configuration>
+
   <appSettings>
-    <add key="aspnet:UseTaskFriendlySynchronizationContext" value="true"/>
+    <add key="aspnet:UseTaskFriendlySynchronizationContext" value="true" />
   </appSettings>
   <system.web>
-    <compilation debug="false" targetFramework="4.7.2">
-      <assemblies>
-        <add assembly="System.ComponentModel.Composition, Version=4.0.0.0, Culture=neutral, PublicKeyToken=B77A5C561934E089"/>
-      </assemblies>
-    </compilation>
+    <compilation debug="false" targetFramework="4.7.2" />
     <httpRuntime targetFramework="4.7.2"/>
   </system.web>
   <system.serviceModel>
+	  <serviceHostingEnvironment>
+		  <serviceActivations>
+			  <add relativeAddress="Service.svc" service="WCFCalculatorServiceLocal" />
+		  </serviceActivations>
+	  </serviceHostingEnvironment>
     <behaviors>
       <serviceBehaviors>
         <behavior>
@@ -23,9 +25,8 @@
       </serviceBehaviors>
     </behaviors>
     <protocolMapping>
-      <add binding="basicHttpsBinding" scheme="https"/>
+        <add binding="basicHttpsBinding" scheme="https" />
     </protocolMapping>
-    <serviceHostingEnvironment aspNetCompatibilityEnabled="true" multipleSiteBindingsEnabled="true"/>
   </system.serviceModel>
   <system.webServer>
     <modules runAllManagedModulesForAllRequests="true"/>
@@ -35,4 +36,5 @@
       -->
     <directoryBrowse enabled="true"/>
   </system.webServer>
+
 </configuration>
